Create file for middleware
