+++ main.go
@@ -14,12 +14,9 @@ import (
 
 	"github.com/EFForg/starttls-scanner/db"
 	"github.com/EFForg/starttls-scanner/policy"
+
 	"github.com/getsentry/raven-go"
-	"github.com/gorilla/handlers"
 	_ "github.com/joho/godotenv/autoload"
-	"github.com/ulule/limiter"
-	"github.com/ulule/limiter/drivers/middleware/stdlib"
-	"github.com/ulule/limiter/drivers/store/memory"
 )
 
 func validPort(port string) (string, error) {
@@ -27,33 +24,6 @@ func validPort(port string) (string, error) {
 		return "", fmt.Errorf("Given portstring %s is invalid", port)
 	}
 	return fmt.Sprintf(":%s", port), nil
-}
-
-func throttle(period time.Duration, limit int64, f http.Handler) http.Handler {
-	rateLimitStore := memory.NewStore()
-	rate := limiter.Rate{
-		Period: period,
-		Limit:  limit,
-	}
-	rateLimiter := stdlib.NewMiddleware(limiter.New(rateLimitStore, rate),
-		stdlib.WithForwardHeader(true))
-	return rateLimiter.Handler(f)
-}
-
-func recoveryHandler(f http.Handler) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-
-		defer func() {
-			if err, ok := recover().(error); ok {
-				rvalStr := fmt.Sprint(err)
-				packet := raven.NewPacket(rvalStr, raven.NewException(err.(error), raven.GetOrNewStacktrace(err.(error), 2, 3, nil)), raven.NewHttp(r))
-				raven.Capture(packet, nil)
-				w.WriteHeader(http.StatusInternalServerError)
-			}
-		}()
-
-		f.ServeHTTP(w, r)
-	})
 }
 
 func pingHandler(w http.ResponseWriter, r *http.Request) {
@@ -65,16 +35,11 @@ func registerHandlers(api *API, mux *http.ServeMux) http.Handler {
 	mux.HandleFunc("/api/scan", apiWrapper(api.Scan))
 	// Throttle the queue endpoint more aggressively so we don't send junk e-mail.
 	mux.Handle("/api/queue",
-		throttle(time.Hour, 3, http.HandlerFunc(apiWrapper(api.Queue))))
+		throttleHandler(time.Hour, 3, http.HandlerFunc(apiWrapper(api.Queue))))
 	mux.HandleFunc("/api/validate", apiWrapper(api.Validate))
 	mux.HandleFunc("/api/ping", pingHandler)
 
-	originsOk := handlers.AllowedOrigins([]string{os.Getenv("ALLOWED_ORIGINS")})
-
-	handler := recoveryHandler(
-		throttle(time.Minute, 10, handlers.CORS(originsOk)(mux)),
-	)
-	return handlers.LoggingHandler(os.Stdout, handler)
+	return middleware(mux)
 }
 
 // ServePublicEndpoints serves all public HTTP endpoints.
